refactor(chain)!: complete removal of `canonical_iter` module
- Delete entire `canonical_iter.rs` file and its module declaration
- Move `CanonicalReason`, `ObservedIn`, and `CanonicalizationParams` to `canonical_task.rs`
- Update module exports to use `pub use canonical_task::*` instead of selective exports

BREAKING CHANGE: `CanonicalIter` and all its exports are removed
